Refactor modal dialogs and add deposit/withdraw dialogs
Moved modal components to a dedicated dialogs directory and renamed CreateVaultModal to CreateVaultDialog. Added new DepositDialog and WithdrawDialog components for vault actions, refactoring dashboard and vault pages to use these dialogs for improved code organization and maintainability.
